exported generatorFunction to own file
